requirements.txt
litellm
pydantic
rich
loguru
python-dotenv